Correct recursive testing
Travis is starting tests with `go test -v ./...`
which adds arguments to the test funtions. As arguments are parsed by
`config` package, it recieves bad arguments and can not find config
file.

This fix prevents looking for file from argument if we are in
testin environment (TEST_ENV is set).

Signed-off-by: Drasko DRASKOVIC <drasko.draskovic@gmail.com>